Simplify app structure by combining App.tsx and main.tsx into a single file
